requirements.txt
streamlit
langchain
faiss-cpu
pypdf
python-dotenv
openai
tiktoken